auto-switch, refactor
